chore(ci): change strategy regarding gitlab ci base image and npm version
First, gitbook-cli can't be installed locally. It must be installed globally cause else it will install a local obsolete npm version.
Secondly, it seems there is not a single docker image on docker hub that ships with all chrome mandatory libs to execute karma and protractor.
The simplest is to start from a basic node image and install dependencies manually.